requirements_no_torch.txt
numpy
h5py
tqdm
datasets
pyarrow
pelicanfs
zarr
wipac-rest-tools